+++ sdk/storage/Azure.Storage.DataMovement/src/StorageResourceSingle.cs
@@ -134,7 +134,7 @@ namespace Azure.Storage.DataMovement
             StorageResourceSingle sourceResource,
             HttpRange range,
             bool overwrite,
-            long completeLength = 0,
+            long completeLength,
             StorageResourceCopyFromUriOptions options = default,
             CancellationToken cancellationToken = default);